token and other are completed

+++ app/resources/item.py
@@ -1,5 +1,10 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required)
 from models.item import ItemModel
 
 
@@ -12,7 +17,7 @@ class Item(Resource):
         'store_id', type=int, required=True, help="Every item needs a store id. "
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -20,6 +25,7 @@ class Item(Resource):
 
         return{'Message': "Item not Found"}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': 'An item name {} already exist.'.format(name)}, 400
@@ -34,7 +40,12 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin Privilege required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -56,6 +67,11 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-
-        return {"items": [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {"items": [item['name'] for item in items],
+                'message': 'More data avairable if you log in'}, 200
